pyproject.toml
[tool.poetry]
name = "langgraph-multi-agent"
version = "0.1.0"
description = "langgraph multi agent environment"
authors = ["JminJ <jminju254@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
python-dotenv = "^1.0.1"
arxiv = "^2.1.0"
feedparser = "6.0.10"
pdfplumber = "^0.11.0"
beautifulsoup4 = "^4.12.3"
tabula-py = "^2.9.3"
jpype1 = "^1.5.0"
langgraph = "^0.2.0"
langchain = "^0.2.5"
langchain-elasticsearch = "^0.2.2"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.5"
duckduckgo-search = "^6.1.7"
pymupdf = "^1.24.10"
wget = "^3.2"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
